+++ src/types/messages/index.d.ts
@@ -19,6 +19,8 @@ export type IncompleteMessage = {
 	isForwarded: boolean;
 	isInvite: boolean;
 	isDraft: boolean;
+	isScheduled: boolean;
+	autoSendTime?: number;
 	participants?: Array<Participant>;
 	date: number;
 	subject: string;
@@ -58,4 +60,5 @@ export type MailMessage = IncompleteMessage & {
 export type SendMsgParameters = {
 	editorId: string;
 	msg?: MailMessage;
+	prefs?: PrefsType;
 };