+++ src/views/app-view.jsx
@@ -8,6 +8,7 @@ import { Redirect, Switch, Route, useRouteMatch } from 'react-router-dom';
 import { FOLDERS, setAppContext, Spinner, useUserSettings } from '@zextras/carbonio-shell-ui';
 import { Container } from '@zextras/carbonio-design-system';
 import { useSelector } from 'react-redux';
+import { includes } from 'lodash';
 import { selectCurrentFolder } from '../store/conversations-slice';
 
 const LazyFolderView = lazy(() =>
@@ -23,10 +24,12 @@ const AppView = () => {
 	const [count, setCount] = useState(0);
 	const { zimbraPrefGroupMailBy } = useUserSettings().prefs;
 	const currentFolderId = useSelector(selectCurrentFolder);
+
 	const isMessageView = useMemo(
 		() =>
 			zimbraPrefGroupMailBy
-				? zimbraPrefGroupMailBy === 'message' || currentFolderId === FOLDERS.DRAFTS
+				? zimbraPrefGroupMailBy === 'message' ||
+				  includes([FOLDERS.DRAFTS, FOLDERS.TRASH], currentFolderId)
 				: undefined,
 		[currentFolderId, zimbraPrefGroupMailBy]
 	);