+++ src/views/app/detail-panel/edit/parts/text-editor-container.tsx
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-import { useIntegratedComponent } from '@zextras/carbonio-shell-ui';
-import React, { FC, ReactElement, useContext } from 'react';
+import { useIntegratedComponent, useUserSettings } from '@zextras/carbonio-shell-ui';
+import React, { FC, ReactElement, useCallback, useContext, useMemo, useRef, useState } from 'react';
 import { Row, Container, Text } from '@zextras/carbonio-design-system';
 import { Controller } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
@@ -15,11 +15,37 @@ import * as StyledComp from './edit-view-styled-components';
 type PropType = {
 	onDragOverEvent: () => void;
 	draftSavedAt: string;
+	minHeight: number;
 };
-const TextEditorContainer: FC<PropType> = ({ onDragOverEvent, draftSavedAt }) => {
+const TextEditorContainer: FC<PropType> = ({ onDragOverEvent, draftSavedAt, minHeight }) => {
 	const { control, editor, throttledSaveToDraft, updateSubjectField } = useContext(EditViewContext);
 	const [Composer, composerIsAvailable] = useIntegratedComponent('composer');
 	const [t] = useTranslation();
+
+	const { prefs } = useUserSettings();
+
+	const defaultFontFamily = useMemo<string>(
+		// TODO: Once the Typescript branch is merged it can be removed
+		// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+		// @ts-ignore
+		() => prefs?.zimbraPrefHtmlEditorDefaultFontFamily ?? 'sans-serif',
+		[prefs]
+	);
+
+	const [inputValue, setInputValue] = useState(editor?.text ?? ['', '']);
+	const timeoutRef = useRef<null | ReturnType<typeof setTimeout>>(null);
+	const [showStickyTime, setStickyTime] = useState(false);
+
+	const toggleStickyTime = useCallback(() => {
+		clearTimeout(timeoutRef.current as ReturnType<typeof setTimeout>);
+		setStickyTime(false);
+		timeoutRef.current = setTimeout(() => {
+			setStickyTime(true);
+			setTimeout(() => {
+				setStickyTime(false);
+			}, 1500);
+		}, 1500);
+	}, []);
 
 	return (
 		<>
@@ -31,29 +57,22 @@ const TextEditorContainer: FC<PropType> = ({ onDragOverEvent, draftSavedAt }) =>
 					crossAlignment="flex-end"
 				>
 					{editor?.richText && composerIsAvailable ? (
-						<Controller
-							height="fit"
-							name="text"
-							control={control}
-							defaultValue={editor?.text}
-							render={({ onChange, value }): ReactElement => (
-								<Container background="gray6">
-									<StyledComp.EditorWrapper>
-										<Composer
-											// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-											// @ts-ignore
-											value={value[1]}
-											onEditorChange={(ev: Array<string>): void => {
-												updateSubjectField({ text: [ev[0], ev[1]] });
-												throttledSaveToDraft({ text: [ev[0], ev[1]] });
-												onChange([ev[0], ev[1]]);
-											}}
-											onDragOver={onDragOverEvent}
-										/>
-									</StyledComp.EditorWrapper>
-								</Container>
-							)}
-						/>
+						<Container background="gray6" mainAlignment="flex-start" style={{ minHeight }}>
+							<StyledComp.EditorWrapper>
+								<Composer
+									// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+									// @ts-ignore
+									value={inputValue[1]}
+									onEditorChange={(ev: Array<string>): void => {
+										setInputValue(ev);
+										updateSubjectField({ text: ev });
+										throttledSaveToDraft({ text: ev });
+										toggleStickyTime();
+									}}
+									onDragOver={onDragOverEvent}
+								/>
+							</StyledComp.EditorWrapper>
+						</Container>
 					) : (
 						<Controller
 							name="text"
@@ -63,6 +82,7 @@ const TextEditorContainer: FC<PropType> = ({ onDragOverEvent, draftSavedAt }) =>
 								<Container background="gray6" height="fit">
 									<StyledComp.TextArea
 										value={value[0]}
+										style={{ fontFamily: defaultFontFamily }}
 										onChange={(ev): void => {
 											// eslint-disable-next-line no-param-reassign
 											ev.target.style.height = 'auto';
@@ -78,28 +98,30 @@ const TextEditorContainer: FC<PropType> = ({ onDragOverEvent, draftSavedAt }) =>
 											throttledSaveToDraft({ text: data });
 											updateSubjectField({ text: data });
 											onChange(data);
+											toggleStickyTime();
 										}}
 									/>
 								</Container>
 							)}
 						/>
 					)}
-
-					{draftSavedAt && (
-						<StyledComp.StickyTime>
-							<Row
-								crossAlignment="flex-end"
-								background="gray5"
-								padding={{ vertical: 'medium', horizontal: 'large' }}
-							>
-								<Text size="extrasmall" color="secondary">
-									{t('message.email_saved_at', {
-										time: draftSavedAt,
-										defaultValue: 'Email saved as draft at {{time}}'
-									})}
-								</Text>
-							</Row>
-						</StyledComp.StickyTime>
+					{draftSavedAt && showStickyTime && (
+						<StyledComp.StickyTimeContainer>
+							<StyledComp.StickyTime>
+								<Row
+									crossAlignment="flex-end"
+									background="gray5"
+									padding={{ vertical: 'medium', horizontal: 'large' }}
+								>
+									<Text size="extrasmall" color="secondary">
+										{t('message.email_saved_at', {
+											time: draftSavedAt,
+											defaultValue: 'Email saved as draft at {{time}}'
+										})}
+									</Text>
+								</Row>
+							</StyledComp.StickyTime>
+						</StyledComp.StickyTimeContainer>
 					)}
 					<StyledComp.Divider />
 				</Container>