+++ src/views/app/detail-panel/preview/attachments-block.jsx
@@ -6,7 +6,7 @@
 import React, { useCallback, useMemo, useRef, useState, useContext } from 'react';
 import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
-import { filter, find, map, reduce, uniqBy } from 'lodash';
+import { filter, find, includes, map, reduce, uniqBy } from 'lodash';
 import {
 	Container,
 	Icon,
@@ -39,11 +39,28 @@ function findAttachments(parts, acc) {
 	);
 }
 
-function getAttachmentsLink(messageId, messageSubject, attachments) {
+function getAttachmentsDownloadLink(messageId, messageSubject, attachments) {
 	if (attachments.length > 1) {
 		return `/service/home/~/?auth=co&id=${messageId}&filename=${messageSubject}&charset=UTF-8&part=${attachments.join(
 			','
 		)}&disp=a&fmt=zip`;
+	}
+	return `/service/home/~/?auth=co&id=${messageId}&part=${attachments.join(',')}&disp=a`;
+}
+
+function getAttachmentsLink(messageId, messageSubject, attachments, attachmentType) {
+	if (!messageId.includes(':')) {
+		if (attachments.length > 1) {
+			return `/service/home/~/?auth=co&id=${messageId}&filename=${messageSubject}&charset=UTF-8&part=${attachments.join(
+				','
+			)}&disp=a&fmt=zip`;
+		}
+		if (includes(['image/png', 'image/jpeg', 'image/jpg'], attachmentType)) {
+			return `/service/preview/image/${messageId}/${attachments.join(',')}/0x0/?quality=high`;
+		}
+		if (includes(['application/pdf'], attachmentType)) {
+			return `/service/preview/pdf/${messageId}/${attachments.join(',')}/?first_page=1`;
+		}
 	}
 	return `/service/home/~/?auth=co&id=${messageId}&part=${attachments.join(',')}&disp=a`;
 }
@@ -90,7 +107,7 @@ const AttachmentExtension = styled(Text)`
 	margin-right: ${({ theme }) => theme.sizes.padding.small};
 `;
 
-function Attachment({ filename, size, link, message, part, iconColors, att }) {
+function Attachment({ filename, size, link, downloadlink, message, part, iconColors, att }) {
 	const { createPreview } = useContext(PreviewsManagerContext);
 	const extension = getFileExtension(att);
 	const sizeLabel = useMemo(() => humanFileSize(size), [size]);
@@ -255,7 +272,7 @@ function Attachment({ filename, size, link, message, part, iconColors, att }) {
 				target="_blank"
 				href={`/service/home/~/?auth=co&id=${message.id}&part=${part}`}
 			/>
-			<AttachmentLink ref={inputRef} rel="noopener" target="_blank" href={link} />
+			<AttachmentLink ref={inputRef} rel="noopener" target="_blank" href={downloadlink} />
 		</AttachmentContainer>
 	);
 }
@@ -276,7 +293,7 @@ export default function AttachmentsBlock({ message }) {
 	const attachmentsParts = useMemo(() => map(attachments, 'name'), [attachments]);
 	const theme = useTheme();
 	const actionsDownloadLink = useMemo(
-		() => getAttachmentsLink(message.id, message.subject, attachmentsParts),
+		() => getAttachmentsDownloadLink(message.id, message.subject, attachmentsParts),
 		[message, attachmentsParts]
 	);
 	const removeAttachments = useCallback(() => removeAttachments(), []);
@@ -375,7 +392,8 @@ export default function AttachmentsBlock({ message }) {
 							key={`att-${att.filename}-${index}`}
 							filename={att.filename}
 							size={att.size}
-							link={getAttachmentsLink(message.id, message.subject, [att.name])}
+							link={getAttachmentsLink(message.id, message.subject, [att.name], att.contentType)}
+							downloadlink={getAttachmentsDownloadLink(message.id, message.subject, [att.name])}
 							message={message}
 							part={att.name}
 							iconColors={iconColors}
@@ -424,7 +442,7 @@ export default function AttachmentsBlock({ message }) {
 						{t('label.download', {
 							count: attachmentsCount,
 							defaultValue: 'Download',
-							defaultValue_plural: 'Downloads'
+							defaultValue_plural: 'Download all'
 						})}
 					</Link>
 					{isUploadIntegrationAvailable && (