+++ src/views/search/parts/use-disable-hooks.js
@@ -6,15 +6,22 @@
 import { isEqualWith } from 'lodash';
 import { useMemo } from 'react';
 
-export const useDisabled = ({ queryToBe, query }) =>
+export const useDisabled = ({
+	queryToBe,
+	query,
+	isSharedFolderIncluded,
+	isSharedFolderIncludedTobe
+}) =>
 	useMemo(
 		() =>
-			isEqualWith(queryToBe, query, (newQuery, currentQuery) => {
-				if (newQuery.length === 0 && currentQuery.length === 0) return true;
-				if (newQuery.length !== currentQuery.length) return false;
-				return newQuery[0].value === currentQuery[0].value;
-			}),
-		[query, queryToBe]
+			isSharedFolderIncluded !== isSharedFolderIncludedTobe
+				? false
+				: isEqualWith(queryToBe, query, (newQuery, currentQuery) => {
+						if (newQuery.length === 0 && currentQuery.length === 0) return true;
+						if (newQuery.length !== currentQuery.length) return false;
+						return newQuery[0].value === currentQuery[0].value;
+				  }),
+		[isSharedFolderIncluded, isSharedFolderIncludedTobe, query, queryToBe]
 	);
 
 export const useSecondaryDisabled = ({