+++ src/views/settings/filters/parts/new-filter-action-conditions.tsx
@@ -6,23 +6,11 @@
 import React, { FC, ReactElement, useMemo, useCallback, useState } from 'react';
 import { Container, Text, Icon, Row, Padding } from '@zextras/carbonio-design-system';
 import { map } from 'lodash';
-import styled from 'styled-components';
+
 import { Folder, getTags, ZIMBRA_STANDARD_COLORS } from '@zextras/carbonio-shell-ui';
 import Heading from '../../components/settings-heading';
 import MoveToFolderModal from './move-to-folder-modal';
 import FilterActionRows from './filter-action-rows';
-
-export const StyledIcon = styled(Icon)`
-	border: 1px solid
-		${({ theme, disabled, color }): string =>
-			disabled ? theme.palette.gray2.regular : theme.palette[color].regular};
-	border-radius: 4px;
-	width: 32px;
-	height: 32px;
-	max-width: 32px;
-	max-height: 32px;
-	padding: 4px;
-`;
 
 type ComponentProps = any;
 
@@ -75,12 +63,7 @@ const FilterActionConditions: FC<ComponentProps> = ({ compProps }): ReactElement
 			<Container padding={{ top: 'medium' }} crossAlignment="flex-start" mainAlignment="flex-start">
 				<Heading title={t('settings.actions', 'Actions')} size="medium" />
 				<Text>{t('settings.perform_following_action', 'Perform the following actions:')}</Text>
-				<Container
-					maxHeight="180px"
-					style={{ overflow: 'auto' }}
-					padding={{ top: 'small' }}
-					mainAlignment="flex-start"
-				>
+				<Container padding={{ top: 'small' }} mainAlignment="flex-start">
 					{map(tempActions, (tempAction, index) => (
 						<FilterActionRows
 							key={tempAction.id}