+++ src/views/settings/subsections.tsx
@@ -8,7 +8,7 @@ import { TFunction } from 'react-i18next';
 // this ignore can be removed after the next rc
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
-import { SettingsSubSection } from '@zextras/carbonio-shell-ui';
+import { getBridgedFunctions, SettingsSubSection } from '@zextras/carbonio-shell-ui';
 
 export const displayingMessagesSubSection = (t: TFunction): SettingsSubSection => ({
 	label: t('settings.label.display_messages', 'Displaying Messages'),
@@ -17,6 +17,11 @@ export const displayingMessagesSubSection = (t: TFunction): SettingsSubSection =
 export const receivingMessagesSubSection = (t: TFunction): SettingsSubSection => ({
 	label: t('label.receive_message', 'Receiving Messages'),
 	id: 'receiving_messages'
+});
+
+export const domainWhitelistSubSection = (t: TFunction): SettingsSubSection => ({
+	label: t('label.trusted_addresses', 'Trusted addresses'),
+	id: 'trusted_addresses'
 });
 export const signaturesSubSection = (t: TFunction): SettingsSubSection => ({
 	label: t('signatures.signature_heading', 'Signatures'),
@@ -36,5 +41,11 @@ export const getSettingsSubSections = (t: TFunction): Array<SettingsSubSection> 
 	receivingMessagesSubSection(t),
 	signaturesSubSection(t),
 	setDefaultSignaturesSubSection(t),
-	filtersSubSection(t)
+	filtersSubSection(t),
+	domainWhitelistSubSection(t)
 ];
+
+export const composingMsgSubSection = (): SettingsSubSection => ({
+	label: getBridgedFunctions()?.t('labels.composing_messages', 'Composing Messages'),
+	id: 'compose'
+});